[classpath products] introduce ResolvedJar and M2Coordinate and use them for improved exclude handling
With this, now excludes of resolved jars will be decided by comparing the org/name of the jar instead of doing path comparisons. This is cleaner, and it'll make it easier to change how excludes works in a follow on change I'm planning.

I based the naming / grouping of data on the entities described in https://docs.google.com/document/d/1sEMXUmj7v-YCBZ_wHLpCFjkHOeWjsc1NR1hRIJ9uCZ8/edit#heading=h.8mden3kdj1co

My thought is that by stepping into this in a constrained way, we can start migrating towards a model that looks like that internally. It should also make it easier do the next thing.