Allow specification of kwargs that are not currently known.
This is useful when preparing BUILD files for versions of pants
that aren't ready to be used yet.

From discussion on pants-devel: Ignoring known unconsumed target
  kwargs.

This change was controversial/distasteful to many on the
pants-devel discussion. I think this is a pretty small change that
doesn't add too much complexity, but if reviewers disagree, we can
just use this patch internally.