Explicitly re-register subsystem options on all per-optionable scopes.
Previously we registered all subsystem options only on their global
instance scope, and relied on "recursive=True" to allow per-optionable
instances to have different option values.  However this is an abuse
of recursive=True, which was really intended only for a small set of
global options, such as log level, to be overridden in specific scopes.

An optionable-specific subsystem exists entirely so that its options
can be set differently than the global instance, so we shouldn't
require recursive=True for that to work, we should just register
everything on every necessary scope.