Hack around a few Distribution issues in py tests.
Distribution gained an undeclared direct dependency on the JVM
subsystem, fouling pyhton tests exercising the antlr builder when run
in isolation.  This change hacks in the deps - which point out that the
JVM Subsystem needs to move out of the jvm backend - and props up a JVM
Subsystem by hand in the 2 affected tests.