Ensure stdlib check uses normalized paths.
Previously, it was observed that the python __file__ might point to a
symlinked path whereas the `sysconfig.get_python_lib(standard_lib=1)`
path might be to the real path.  This led to mis-categorization of
stdlib imports as third party.  This fix ensures all paths are realpaths
to remove any ambiguity.

Also remove the unused PLAT_SPECIFIC_PATH.