Fix PythonThriftBuilder to operate on 1 target.
Previously it operated on the closure of thrift files depended on by a
single PythonThriftLibrary and this trampled on work both PythonChroot
and SetupPy did to walk the closure and link together CodeGenerator
outputs.  This brings PythonThriftBuilder into line PythonAntlrBuilder
and fixes 2 bugs.

Tests are added that fail before the fix and pass after.