Fixup Fixup `TargetTest.test_unknown_kwargs`.
Elevate a `Target.type_alias` property that works for any target and
faithfully reflects the target's type to the target author.  For BUILD
file author's, this is the type alias used in the BUILD file, for
test writers and other code synthesizing targets, this is the target
class name.

Also add some new and missing docs.