Make nailgun_client error when the client socket is closed. [BUGFIX]
Previously, if the client socket was closed gracefully on
the other side (by the nailgun) while at the start or middle
of attempting to read a message from the socket, the client
would go into a hard loop attempting to do a blocking read
on the socket.  But when the socket has been closed, that
blocking read returns immediately with no bytes to indicate
that the socket is closed.

Testing Done:
Added tests to exercise the two new exception types and the
various interesting points in the message byte stream where
the server could terminate the socket.

CI is green: https://travis-ci.org/pantsbuild/pants/builds/77937754

Bugs closed: 2091

Reviewed at https://rbcommons.com/s/twitter/r/2727/