Ensure pipeline failuires fail CI.
We use pipelines to collect filtered lists of targets for
several CI steps.  Previously these could fail when BUILD
files had sytax or other errors, but the CI run would proceed
with no targets.  This change turns on pipefail and lifts up
target collection steps so they are AND'ed with the steps
that use them such that pipefail actually triggers an overall
failure in those steps.