Introduce anonymous targets built by macros.
This change introduces a new `BuildFileAliases` slot for "anonymous
targets" that should be exposed only to context aware object factories,
aka: BUILD file macros.  This allows a target author to be more
restrictive with a target's arguments.  Two cases are immediately
important:

1. Enforcing 1:1:1
   This requires taking away the ability of the BUILD file author to
   specify a target name to ensure each BUILD file can have only 1
   target.  Since name is integral to Target capture, only a macro can
   do this.
2. Enforcing no dependencies
   The need to do this at all should go away when BuildGraph handling of
   non-Target Addressables is supported; but, until then, dependencies
   is another low-level property of Targets that cannot otherwise be
   blocked from population by BUILD file authors.

In the course of adding support for anonymous targets, the limits of the
existing system of carrying around type alias metadata in
BuildFileAddress were reached, and so this attribute was re-plumbed
through the Addressable proxies and into Target directly, allowing
macros to specify the alias for anonymous targets they create.

Finally, the BuildFileTypeFactory was lifted up as a small interface to
abstract the role both macros and target proxies serve in SourceRoot
specification and an explicit Macro type was introduced to allow for
macros to back restricted source roots.