Kill PANTS_VERBOSE and PANTS_BUILD_ROOT.
The PANTS_VERBOSE env var was obsolete and just needed a logger plumbed
and the PANTS_BUILD_ROOT was unused and unuseable since code elsewhere
in Config does `os.path.join(get_buildroot(), 'pants.ini')` which
assumes no PANTS_BUILD_ROOT env var override took place (because the
default way to establish the build root is in fact the inverse of the
path join!).

Docs and messages are all updated to kill mention of these 2 dead env
vars and a few extra doc lies are removed or fixed.