Eliminated duplicate jar entries caused by consecutive slashes.
We noticed this duplication in some of our internal jar builds. It
mostly doesn't cause any problems, but was breaking jarjar when it
tried to shade them (for rb 2754).

Despite only having one version of the class most of the way
through the jar building pipeline, duplicate classes were
artificially created due to a bug in the path calculation for the
jar entry.

Specifically, we create the final jar entry path by calling 'join'
on the path separator ('/') over an iterable of path components,
however it appears there are some cases where the path components
may contain trailing slashes.

Internally, this resulted in .jar files with entries that looked
like:

    com/squareup/foo/bar/app/package/File.class
    com//squareup/foo/bar/app/package/File.class
    com/squareup//foo/bar/app/package/File.class
    com/squareup/foo//bar/app/package/File.class
    com/squareup/foo/bar//app/package/File.class
    com/squareup/foo/bar/app//package/File.class
    com/squareup/foo/bar/app/package//File.class

This is a very simple (but subtle) bug, and the fix is similarly
simple.