Re-work REPL mutual exclusion.
The old method required modifying Target which was untenable for
plugins.  A new approach is taken by introducing a MutexTaskMixin that
supports forming groups of tasks that all provide the same mutually
exclusive service.  The functionality is the same as that provided by
the previous pairing of `Target.lang_discriminator` and
`Task.require_homogeneous_targets`, but with a more flexible version of
`Target.lang_discriminator` that allows any installed tasks with a
shared MutexTaskMixin to form a mutual exclusion group in an
uncoordinated manner (save from mixing in the shared base mutex mixin).

The `ReplTaskMixin` serves as the realization of this concept and both
`PythonRepl` and `ScalaRepl` tasks and tests are retrofitted to use it.

This paves the way for the node contrib plugin to provide its own
`NodeRepl` in a follow-up RB.