Add support for `repl.node`.
This builds off of the new ReplTaskMixin and NodeRepl itself is a small
task.

To support NodeRepl well, a few adjustments are made:
1. MutexTaskMixin is given better failure output when identifying
   mutually incompatible targets.
2. NpmResolve is updated to support resolving remote modules directly.
3. The NodePaths product if broken out of NpmResolve to present a more
   neutral data type and not explicitly couple NpmResolve to NodeRepl.

An example node 3rdparty dep is added to support some integration tests
and the existing NodeTask and NpmResolve tests are expanded to cover the
new functionality added to the respective classes.