Add support for @fromfile option values.
Options can now be registered as supporting fromfiles by passing
`fromfile=True`.  When so registered, if the option value starts with
the '@' character, the rest of the value is interpreted as a file path
and the option value is read from that file.

This will open up support for large configuration values being stored
outside of `pants.ini` and in dedicated configuration files on disk. This
facility will make most sense for large string messages, long lists, and
big dictionaries.