Accept both Target types an TargetMacro.Factory instances in BuildFileAliases.targets.
Since both of these are factories for Targets, though at different layers in
the Target hydration stack, allow both to be registered as "targets".  This
allows for the two concepts to potentialy merge later, ie: Target might have
an defaulted factory method that subtypes could override to provide macro-like
behavior.