Replace eval use with a parse_literal util.
This locks down and centralizes all of pants uses of eval outside BUILD
file parsing in a central utility.  Config and custom_types are updated
to use this and tests adjusted.  In particular, test_options is ported
to use a real Config instead of a FakeConfig - which behaved differently
than real Config and was causing fake - aka test code induced - issues
with the new parser.