Make BootstrapJvmTools fail fast on a bad user-specified address.
Previously, if you specified a target for a jvm-tool that pants
failed to resolve, if the jvm-tool had a default classpath, pants
would simply inject a synthetic JarLibrary with the same name that
used the default classpath.

This resulted in pants pretending to use the user-specified target
that it couldn't find, but silently it was using the default.

This caused problems for square internally when we tried to specify
a different version of the wire-compiler with an (unwittingly)
malformatted spec. The build continued along merrily and subtly
pulled in the wrong version of wire, only blowing up when it tried
to compile a target that used features not available in pants's
default version of wire.

Example output after this change:

    ~/Src/Pants ./pants --gen-wire-wire-compiler="this-target-definitely-does-not-exist" clean-all compile examples/src/wire::
    INFO] Detected git repository at /Users/gmalmquist/Src/Pants on branch gmalmquist/missing-jvm-tool-options-fail-loudly

    17:41:55 00:00 [main]
                   (To run a reporting server: ./pants server)
    17:41:55 00:00   [bootstrap]
    17:41:55 00:00   [setup]
    17:41:55 00:00     [parse]
    FAILURE: Failed to resolve target for tool: this-target-definitely-does-not-exist. This target was obtained from
    option wire-compiler in scope gen.wire.

    Make sure you didn't make a typo in the tool's address. You specified that the
    tool should use the target found at "this-target-definitely-does-not-exist".

    This target has a default classpath configured, so you can simply remove:
      [gen.wire]
      wire-compiler: this-target-definitely-does-not-exist
    from pants.ini (or any other config file) to use the default tool.

    The default classpath is: com.squareup.wire-wire-compiler-1.6.0

    Note that tool target addresses in pants.ini should be specified *without* quotes.

    17:41:55 00:00   [complete]
                   FAILURE