A nice timing stats report.
Consists of line charts showing aggregate workunit timings by date,
and a drilldown table with further detail per sub-workunit.

Note that, obviously, the timings likely change dramatically day-to-day,
depending on how many builds the user runs, and what they do.
So it's hard to come up with any generally useful report idiom.
This report represents how much time a user is spending in various
parts of Pants during a day, which I hope is a rough indicator of sources
of user dissatisfaction (given that we assume that the ideal time for
any build operation is zero), and can inform optimization decisions.

Also makes the pants server serve a favicon, as the JS console errors
on lack of favicon were driving me nuts, and, amazingly, there's no reliable
way for a page to tell a browser to stop requesting favicons, other than
to provide one...