Improve error message when Python headers are missing
You can find a detailed explanation for this change in issue #2161.

If you build pants on systems without Python development headers then `pip` will
fail with a rather uninformative error message.  This change improves the
error message to suggest a possible fix in order to improve the end user
experience.