Kill 'ivy_resolve_symlink_map' and 'ivy_cache_dir' products.
The 'ivy_cache_dir' product was unused within pants and anyone in need
of this info can consult the IvySubsystem.

The 'ivy_resolve_symlink_map' product was unused except internally to
the `IvyTaskMixin`/`IvyResolve` - switch to a normal return parameter
and kill the one unused task dependency on the product.  As a result,
simplify symlink handling which no longer is multi-pass as is made clear
by the product collapse.