Fix `NailgunExecutor` to support more than one connect attempt
In 73e25e74, the `WAIT_INTERVAL` field of `ProcessManager` was renamed to
`WAIT_INTERVAL_SEC`, except for one location in `NailgunExecutor` which broke as
a result.  This causes the `try_connect` method to fail if one connection
attempt fails, regardless of what the user supplies as the
`--nailgun-connect-attempts` flag on the command line.