Kill 'ivy_jar_products'.
Previously this was only used by the IDE focused tasks: `Export` and
`IdeGen`.  Enhance ClasspathProducts by exposing ClasspathEntries and
enhance the ArtifactClasspathEntry to carry the ivy cache path of the
artifact in addition to the pants path.

Also fix up the core ivy.xml template to successfully pull extra confs
(sources and/or javadoc) when custom classifiers have been requested.
This just required using the <include/> directive for the extra confs.
Uncomment a few tests that exercise this.