Fix `linecount` estimator for `dep-usage` goal
In 8c82cf54, the `file_line_count` function was renamed to `line_count`, but
one of the occurrences was renamed to `linecount` by mistake.  This causes the
`dep-usage` goal to fail when provided with the `linecount` estimator.

This change adds a new integration test for the `dep-usage` goal which failed
before the change and succeeds after.