Kill 'jvm_build_tools_classpath_callbacks' deps.
Only `JvmToolMixin` depends on this product, so isolate the product
dependency to it.  This necessitated some plumbing since `JvmToolMixin`
is not itself a task with a `prepare method`.  This also required
relaxing `RoundEngine` validation of product dep cycles.  It now trusts
you know what you're doing as the author of a task that produces
products it also depends on.