Migrate `JvmBinaryTask` to 'compile_classpath'.
The 'jar_dependencies' is slated for imminent removal and the
`ArtifactClasspathEntry`s exposed via the 'compile_classpath' product
provide richer data that is used to ensure unique jar paths in
`BundleCreate`.  In addition, `JvmBinaryTask`s handling of
`JvmBinary.deploy_excludes` now leverage the same exclude logic used in
forming classpaths via the help of a new `ClasspathProducts.copy` method
that makes it safe for `JvmBinaryTask` to snapshot the classpath and add
in additional deply-only excludes for forming deploy binaries and
bundles.

The use-case of grabbing external artifact only classpaths from
`ClasspathProducts` is formalized in a new API and tests are added for
all the new products behaviors.