Add support for snapshotting `ClasspathProducts`.
This is part of the story in supporting proper transitive
`JvmBinary.deploy_excludes` in `JvmBinaryTask` derived classes and
in-general sets up users of these products to take a snapshot and then
build upon the products without pollution.

Tests are added to verify the copies used in snapshotting are
sufficiently divorced from their originals.