Transition `IvyImports` to 'compile_classpath'.
This kills the last user of the 'jar_dependencies' product and, as a
result, the last extra product of `IvyResolve` - all resolve consumers
now use the 'compile_classpath' `ClasspathProducts`.

To support this, the majority of `IvyResolve.execute` is pulled up into
`IvyTaskMixin.resolve` which now handles populate of `ClasspathProducts`
from the lower-level `ivy_resolve`.  Much cleanup in IvyTaskMixin is now
tractable, but left for follow up changes.

The `IvyImports` product is hardened into a custom type and consumers
are updated to use this type and remove some prior hacks no longer
needed.

The 'jar_dependencies' product was the only one to use the `Products`
required predicate mechanism, so that mechanism is removed to bring
`ProductMapping` products in-line with data products.  A subsequent
change can now merge these two types of product into one with identical
handling.