Don't modify globs excludes argument value.
Ideally, users shouldn't be using variables in BUILD files, but if they do, we shouldn't be modifying their values.

The error message that comes up is like this when the excludes value is reused as a sources list:

                       File ".../src/python/pants/base/target.py", line 545, in create_sources_field
                         assert_list(sources, key_arg=key_arg)
                       File ".../src/python/pants/base/validation.py", line 49, in assert_list
                         .format(key_msg, expected_type, e, e.__class__))
                     ValueError: In key 'sources': Expected a list containing values of type (<type 'basestring'>,), instead got a value ['fleem.java'] of <type 'list'>

Testing Done:
Wrote regression test, saw that it failed. Updated implementation and made it pass.

Bugs closed: 2213

Reviewed at https://rbcommons.com/s/twitter/r/2841/