Include resolved jar versions in compile fingerprints; ensure coordinates match artifacts.
We've run into issues recently where a different jar version resolution can result in different compiled outputs. This adds the resolved coordinates into the fingerprint for JVM compile artifacts. By doing that we can ensure that cache entries for compile artifacts will be invalidated if the resolved jar versions change.

Additionally, while working on this I discovered that the revision assigned to a particular classpath artifact is the requested one and not the final one. I added a unversioned -> versioned table to IvyInfo to ensure that the resolved versions are used in all cases.