Fix a protocol bug in `GopkgInFetcher` for v0's.
Previously the `GopkgInFetcher` always fetched the 'master' branch for
v0 packages.  The spec only suggests 'master' as the fallback for v0's
when no v0 matching branch or tag can be found, see the 'Version Zero'
section here:
  http://labix.org/gopkg.in

This change fixes up the fetcher to meet the spec and adds tests for
the v0 cases.