Define local variable in finally block of nailgun_executor
The sock variable was not defined upon failure. This had the
chance of causing zombie nialguns as the socket could be created
but never handled if sock_connect_ex returned non-zero.

The else block in nailgun_client.try_connect was changed to return
sock.close() which closes the socket upin failure while still
returning None to nailgun_executor.

This is intended to maintain the asynch connects.