Fixup release dry runs; use isolated plugin cache.
Previously release dry runs used the global plugin cache.  This was
problematic since pants versions are static in between releases, and so
installation and use of plugins in release dry run tests would use the
last released version of a plugin instead of the version just burned to
`dist/`.  Combined with API changes across plugin boundaries since the
last release, this could lead to spurious failures.

Fix this by adding an advanced `--plugin-cache-dir` option and use this
in the release dry run pointing to a new empty temporary directory.