Introducing the notion of required and optional extra ivy confs.
IdeGen now resolves its own jars intead of depending on the
IvyResolve task.

We don't want ivy to fail if javadocs and sources cannot be
resolved for the ide step, so this allows us to indicate that we
wan't confs if they are available, but we don't need to fail if
we can't find them.

This involved some special logic for artifacts that specify a url,
because the syntax becomes a bit different.