Stopped ivy from failing if an artifact has a url specified.
Previously ivy would fail on jar dependencies with a url
set for the idea and export goals when javadocs and sources
were set to resolve, because ivy would be unable to locate
sources and javadocs for the artifact.

This change disables javadocs and sources for dependencies
that have an artifact that specifies a url.

IdeGen and Export now resolve their own jars intead of depending
on the IvyResolve task.