Allow resolving buildcache hosts via a REST service
Today buildcache urls are statically configured in pants. In Twitter case we
use a centralized load balancer, that becomes a bottleneck with the recent
monobuild launch. Deploying a local proxy is a possibility but not quite ready
yet because we have heterogeneous deployment environments besides aurora,
jenkins for example.

This CL adds an option --resolver to CacheSetup, which point to an endpoint
that supplies buildcache urls. The CL provides a RESTful implementation of the
resolver. --resolver is optional when it is not supplied, today's behavior
remains unchanged.

Testing Done:
https://travis-ci.org/peiyuwang/pants/builds/80303989

Bugs closed: 2171, 2219

Reviewed at https://rbcommons.com/s/twitter/r/2815/