Remove artifacts from JarDependency; Kill IvyArtifact.
The `IvyArtifact` and its collection inside `JarDependency` objects is
killed in favor of declaring 1 jar dependency per needed artifact.  This
was already the way to declare custom classifier dependencies, but the
vestigal artifacts system was engaged prior to this change.  Now
artifact attributes are rolled up into `JarDependency` and only the ivy
resolution code knows to aggregate multiple jars for the same coordinate
with different classifiers into 1 ivy dependency.  The user-facing model
shows no signs of this implementation detail.

The only remaining evidence of ivy in the BUILD UI for jars is the
`type_` paramater which is now deprecated in favor of `ext`.

Tests are updated and `JarPublish` is simplified on the ivy end where
the ivy.xml need not record external dependencies since it just drives
the publication machinery for poms and various jars and other artifacts.

Testing Done:
CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/82093645

Bugs closed: 1410, 2235

Reviewed at https://rbcommons.com/s/twitter/r/2858/