Upstream the scala_jar and scala_artifact helpers to suffix the scala platform version.
Add helpers for suffixing the scala platform version on libraries. We've been using this internally to ease scala upgrades by centralizing the platform version. We could of course do much, much more with this, so consider this revision a strawman.

- Add scala_jar for use with remote dependencies
- Add scala_artifact for use in provides clauses

Most likely what this should do instead would be to define classpaths for each platform, with the major version as a key.

Testing Done:
https://github.com/pantsbuild/pants/pull/2273

Reviewed at https://rbcommons.com/s/twitter/r/2891/