Stop the execution_graph from infinitely adding downstream failures.
We discovered that certain large build graphs were stalling
and paging all available memory into a python process. Upon
dumping the python heap we discovered that there the execution
graph was unboundedly adding failed jobs tuples to the
finished_queue.

This patch merely checks to see if the job is already marked
as done before inserting it into the finished_queue. The
root cause of this behavior is likely undiscovered. One guess
was a cycle in the build graph but the execution_graph tests
supposedly test for that, and pants itself should detect a cycle
in the build graph.