Ensure execution graph cancellation only happens once per job
Before, the execution graph wouldn't check to see if a job had already been canceled before canceling it. This meant that even if that job was a dependee of another canceled or failed job, it would still end up in the queue. The result is that the queue could grow to be very large with deep graphs, containing mostly duplicate cancelations.

This patch changes the cancellation handling so that the status table is checked for each dependee before canceling it. It also marks them as queued so that the status table knows that they are in progress.