Seperate out AddressMapper from Graph.
This restricts responsibility of Graph to just linking addressables.
Currently Graph links via inlining fully hydrated objects, but it or an
alternative could also traverse to addressables on demand.

With AddressMapper extracted, an interface for BUILD file invalidation
comes to the fore and allows straight-forward integration of in
invalidation system that watches the build root for BUILD file
additions, removals and modifications.