Simplify cobertura source paths for reporting
The existing approach to supplying cobertura reporting with access to coverage target source is somewhat complicated, largely with the intention of fooling cobertura into finding scala classes. Cobertura is not really designed to cover scala in the first place, so this seems fairly pointless. Removing that in favor of a much simpler approach that just passes the traget_base of relevant targets.

Testing Done:
Manually tested on a couple of targets, and added validation to an existing integration test that looks for the annotated source file html and verifies it contains some src.

Bugs closed: 2300

Reviewed at https://rbcommons.com/s/twitter/r/2918/