Bump the default ivy bootstrap jar to 2.4.0.
This has been the stable release for almost a year and pantsbuild/pants
itself needs this version to properly handle the ANDROID_HOME env var
detection it uses in ivysetting.xml.

Previously, the bootstrap was 2.3.0 and the final used by
pantsbuild/pants was 2.4.0 + extras.  For any workspace with a cached
final, the ivysettings.xml would work, but for workspaces w/o an ivy
bootstrapped, the ivysettings.xml would fail.  This change gets
pantsbuild/pants ivysettings.xml working no matter the bootstrap phase.