Allow six to float a bit.
We use an extra unittest2 requirement in the `test.pytest` task
and it floats six >=1.4.  Our pin conflicts with the float.  The
conflict appears to be a resolver bug in pants or pex that needs
to be investigated and fixed (see:
https://github.com/pantsbuild/pex/issues/167); this fix is just a
workaround to buy time for that fix.

Testing Done:
CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/84148668

Bugs closed: 2340

Reviewed at https://rbcommons.com/s/twitter/r/2942/