Support a non-inlined lazy resolve mode in Graph.
The fundamental restructuring is moving from Addressed values to a
class-housed schema of addressable fields powered by a @property-like
`AddressableDescriptor` with write-once semantics and that can contain
`Resolvable` values in addition to opaque address strings and concrete
values.  This move reduces the memory requirements of parsing objects,
they are now just an object with a single `_kwargs` dict of unadorned
string address values in the addressable field case.

With the `Resolvable` type plumbed into addressable schemas, both the
json encoder and the Graph gain the ability to materialize or serialize
shallow or inlined graphs bringing the testable space of the noun side
of the experiment up to 3 parsers x 2 graph modes.