Support for scanning addresses, implement `list`.
This also introduces a `legacy_python_callbacks_parser` that can work on
most legacy BuildFileAliases with proper configuration.  This is used to
create a self-contained list command that can be used in legacy BUILD
trees (possibly with some `--spec-excludes` to look past trick macros)
to start perf testing aspects of this experimental parsing system
against the 'legacy' (current) system.