Switch all internal code to use the new source roots mechanism.
In particular, have tests use standard source root names that
match the default patterns, so we don't need to explicitly register
source roots in tests.

Required a couple of hacks to set up the SourceRootConfig
subsystem in a handful of tests that don't use BaseTest.context().
We really need to find a more straightforward way to set up tests.
Hopefully the new engine will remove the need for so much implicit
global state.

The changes are mostly mundane, but there a few files to pay particular
attention to:

- source_root.py
- context.py
- base_test.py
- go_buildgen.py
- maven_layout.py