Minimal Haskell plugin for `pants`
This is a minimum viable subset of a larger plugin that I have been working on
here:

https://github.com/pantsbuild/pants/compare/master...Gabriel439:ggonzalez/haskell-plugin

The easiest way to review this is to read the `contrib/haskell/README.md` that
this is included within this change.  You can also view the rendered `README.md`
at:

https://github.com/Gabriel439/pants/blob/ggonzalez/haskell-plugin/contrib/haskell/README.md

Note that the `README.md` describes the full set of changes that I have already
made on my larger branch, not just the changes within this review board, but
it's probably better that way since you'll have more context for reviewing this
smaller change.

This change provides:

* New `pants` targets for Haskell projects
* Support for `./pants build` on these targets
* Example targets

I wasn't sure whether or not to include the integration tests since @jsirois
asked me to keep this small.  I can expand this review board further to add one
of the integration tests I already wrote if people would like to see that, too.