Introduce the 1st cut at the new engine frontend.
The new `LocalScheduler` implements a means of extracting an
`ExecutionGraph` from `TaskPlanner`'s.  This graph can later be used by
a new engine backend to schedule maximally parallel execution as well as
distributing that execution.

A test is added to exercise the nominal simple codegen case as well as a
few other cases convenient to the example codegen graph.