Handle case where only transitive dependencies have changed during an incremental build
Because of [#2446](https://github.com/pantsbuild/pants/issues/2446), when you add and remove a dep from a target in succession for a task that uses transitive invalidation, it's possible for an incremental compile to occur in the same `results_dir` as the previous attempt. In that case, we were seeing an error (rightly, as it was an unexpected case) because the previous `results_dir` had not been initialized.

- Work around the bug by always initializing `previous_results_dir`, even if it would be the same as the current `results_dir`
- Add a test that doesn't quite reproduce the issue, and expect it to fail

Testing Done:
Fixes:

    compile(src/scala/org/pantsbuild/zinc:zinc) failed: There is no previous_results_dir for: VT(src.scala.org.pantsbuild.zinc.zinc, invalid)

expected to go green:
https://travis-ci.org/pantsbuild/pants/builds/87097649

Bugs closed: 2447

Reviewed at https://rbcommons.com/s/twitter/r/3028/